requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
google-generativeai
python-dotenv
Pillow
Flask-Markdown==0.3